+++ src/firetrace/interface_text.py
@@ -1,0 +1,25 @@
+q_and_a = """        
+        ### **Q: What's the deal with Firetrace? 🔍**\n
+        A: Picture this—it's like having your very own bushfire fortune teller! 
+        🔮 Firetrace is an AI-powered web interface that predicts the severity of bushfire events all across Australia. 
+        🌍 It's armed with projected weather data, BOM weather observatories and NASA's MODIS satellite intel. 
+        It even considers time information to get climate change trends. 🦸‍♂️🔥
+        \n
+
+        ### **Q: What is the inspiration behind Firetrace?** 🤔 \n
+        A: Those bushfires! They cause habitat loss, put many animal species at risk, and drastically impact the economy 😢📉 
+        The series of bushfires that took place in Australia is a clear demonstration of this. 
+        ☝️ So, we put our heads together to conjure up Firetrace—a cutting-edge tool that helps us predict the severity of these fiery events.
+        This will be helpful in making smarter decisions and being prepared to take on the bushfire challenges head-on! 💪🌳
+        \n
+
+        ### **Q: How do I use Firetrace?** 🎉 \n
+        A: It's easy-peasy! 🔥 
+        Simply enter in values for the input boxes below, and you'll unlock access to the hottest predictions, 
+        and uncover the secrets of bushfire severity. 
+        The prediction you receive will be in the form of a percentage - the area covered your bushfire as a percentage of the area covered by the most severe week of the 2011-2012 bushfire season 
+        (Keep in mind that it's not completely right - but it does have reasonable accuracy)
+        \n
+
+        Let's work together to face these fiery challenges to create a safer future! 🚀🌿
+        """
